Fixed a sentence fragment
The context of the rectified sentence implies a general assertion. While the original sentence was telling about a condition without a result.
Original statement: ' AS many of you seen Forbes quotes of the day when you open the Forbes website'  (...THEN what?)
Rectified statement: 'Many of you would have seen the Forbes Quote of the day when you open the Forbes website. '